+++ packages/common/web/lib/imageUtils.ts
@@ -30,34 +30,39 @@ const ceilDimension = (dim: Dimension) => ({ w: Math.ceil(dim.w), h: Math.ceil(d
 
 const createUrl = (
   src: ComponentData['src'],
-  removeUsm?: boolean,
   rw?: number,
   rh?: number,
   rq?: number,
   type = DEFAULT.TYPE,
-  size?: string
+  size?: string,
+  options?: {
+    removeUsm?: boolean;
+    encAutoImageUrls?: boolean;
+  }
 ) => {
   if (type === 'preload') {
-    return createPreloadUrl(src, rw, rh, rq);
+    return createPreloadUrl(src, rw, rh, rq, options?.encAutoImageUrls);
   }
   if (type === 'quailtyPreload') {
-    return createQuailtyPreloadUrl(src, rw, rq, size);
+    return createQuailtyPreloadUrl(src, rw, rq, size, options?.encAutoImageUrls);
   }
-  return createHiResUrl(src, rw, rh, rq, removeUsm);
+  return createHiResUrl(src, rw, rh, rq, options?.removeUsm);
 };
 
 const createPreloadUrl = (
   { file_name: fileName, width: w, height: h }: ComponentData['src'] = {},
   rw = DEFAULT.SIZE,
   rh = DEFAULT.SIZE,
-  rq = DEFAULT.QUALITY
+  rq = DEFAULT.QUALITY,
+  encAutoImageUrls
 ) => {
   if (fileName) {
     const { width, height } = resize(w, h, rw, rh);
     const H = Math.ceil(height); //make sure no sterching will occur
     const W = Math.ceil(width);
     const format = getImageFormat(fileName);
-    return `${WIX_STATIC_URL}/media/${fileName}/v1/fit/w_${W},h_${H},al_c,q_${rq}/file${format}`;
+    const params = `w_${W},h_${H},al_c,q_${rq}${encAutoImageUrls ? ',enc_auto' : ''}`;
+    return `${WIX_STATIC_URL}/media/${fileName}/v1/fit/${params}/file${format}`;
   }
 };
 
@@ -65,16 +70,16 @@ const createQuailtyPreloadUrl = (
   { file_name: fileName, width: w, height: h }: ComponentData['src'] = {},
   rw,
   rq = PRELOAD.QUALITY,
-  size
+  size,
+  encAutoImageUrls
 ) => {
   if (fileName) {
     const width = rw || IMAGE_SIZE[size] || PRELOAD.WIDTH;
     const minW = Math.min(width, w);
     const ratio = h / w;
     const tDim: Dimension = ceilDimension({ w: minW, h: minW * ratio });
-    return `${WIX_STATIC_URL}/media/${fileName}/v1/fit/w_${tDim.w},h_${
-      tDim.h
-    },al_c,q_${rq}/file${getImageFormat(fileName)}`;
+    const params = `w_${tDim.w},h_${tDim.h},al_c,q_${rq}${encAutoImageUrls ? ',enc_auto' : ''}`;
+    return `${WIX_STATIC_URL}/media/${fileName}/v1/fit/${params}/file${getImageFormat(fileName)}`;
   }
   return '';
 };
@@ -113,6 +118,7 @@ const getImageSrc = (
     requiredQuality?: number;
     imageType?: string;
     removeUsm?: boolean;
+    encAutoImageUrls?: boolean;
     size?: string;
   } = {}
 ) => {
@@ -134,12 +140,15 @@ const getImageSrc = (
     } else if (src.file_name) {
       const url = createUrl(
         src,
-        options.removeUsm,
         options.requiredWidth,
         options.requiredHeight,
         options.requiredQuality,
         options.imageType,
-        options.size
+        options.size,
+        {
+          removeUsm: options.removeUsm,
+          encAutoImageUrls: options.encAutoImageUrls,
+        }
       );
       return url;
     }